Cleaned up code in destroy_friendship.rb and DestroyFriendship.rb.
Had to update the Rates.rb helper class. There are endpoints of the
REST API that are rate limited but do NOT appear in the rate limit
information that Twitter provides programmatically!

If Rates.rb encounters such an endpoint it assumes that it can only call
that endpoint 15 times and never refreshes the rate limit information
for that endpoint (there's no point!).